railway_backend_fixed/requirements.txt

fastapi
uvicorn
requests
python-dateutil
